More prototype weapon work: nearing completion
