+++ data/config/prism/prism_weapons_blacklist.csv
@@ -1,0 +1,2 @@
+id
+vass_prototype_s1
