+++ data/strings/tips.json
@@ -2,6 +2,6 @@
 	tips:[
 		"[The Vass] Do not underestimate Vass vessels based on stats alone; their passive time distortion effect turns them considerably faster and deadlier than they appear at first glance.",
 		"[The Vass] While the Families mostly treat each other as equals, or at the very least close in power, the near-mythological leader of Accel stands out among the rest as a figure of power.",
-		"[The Vass] The Families do not kindly to looters and thieves, and will go to great lengths to hide the secrets of their products.",
+		"[The Vass] The Families do not take kindly to looters and thieves, and will go to great lengths to hide the secrets of their products.",
     ],    
 }
