+++ data/weapons/weapon_data.csv
@@ -13,7 +13,10 @@ Colada Emitter,vass_colada,5,,0,800,,300,225,1,45,5,36,0.18,3,FRAGMENTATION,75,,
 Excalibur Emitter,vass_excalibur_launcher,5,,0,800,,1000,750,1,50,0,3,,,FRAGMENTATION,0,,0,7,1,,0,0,0,0,,250,100,6,200,"SYSTEM, SHOW_IN_CODEX, GUIDED_POOR",,,Perturba Tech,,Strike,Fast,None,,,"Disturbs nearby space-time once the projectile destabilizes.",,,,,
 ,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
 #base_prototypes,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
-Perturba Prototype,vass_prototype_s1,5,,0,500,,100,,0,30,10,,,,OTHER,200,,0,4,,,0,0,0,0,,700,,,,,,,Perturba Prototype,,Unknown,Unknown,Unknown,Unknown,Unknown,"%s statistics: no data available. Extremely rough estimates displayed.",Unknown,,,,
+Perturba Prototype,vass_prototype_s1,5,,0,500,,170,,0,30,10,,,,OTHER,200,,0,3.5,,,0,0,0,0,,700,,,,,,,Perturba Prototype,,Unknown,Unknown,Unknown,Unknown,Unknown,"%s statistics: no data available. Extremely rough estimates displayed.",Unknown,,,,
+Perturba Prototype,vass_prototype_s2,5,,0,650,,90,,0,38,10,,,,OTHER,105,,0,1.8,,,0,0,0,0,,600,,,,,,,Perturba Prototype,,Unknown,Unknown,Unknown,Unknown,Unknown,"%s statistics: no data available. Extremely rough estimates displayed.",Unknown,,,,
+Perturba Prototype,vass_prototype_m1,5,,0,600,,400,,0,16,20,,,,OTHER,530,,0,4.5,,,0,0,0,0,,700,,,,,,,Perturba Prototype,,Unknown,Unknown,Unknown,Unknown,Unknown,"%s statistics: no data available. Extremely rough estimates displayed.",Unknown,,,,
+Perturba Prototype,vass_prototype_m2,5,,0,850,,195,,0,19,20,,,,OTHER,220,,0,2.2,,,0,0,0,0,,800,,,,,,,Perturba Prototype,,Unknown,Unknown,Unknown,Unknown,Unknown,"%s statistics: no data available. Extremely rough estimates displayed.",Unknown,,,,
 ,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
 #deco,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
 Makhaira Lights,vass_deco_makhaira_lights,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,SYSTEM,,,,,,,,,,,,,,,
