+++ jars/src/data/scripts/hullmods/VassPeriodicPlating.java
@@ -30,11 +30,6 @@ import org.lwjgl.util.vector.Vector2f;
 public class VassPeriodicPlating extends BaseHullMod {
 	public static final Logger LOGGER = Global.getLogger(VassPeriodicPlating.class);
 
-	@Override
-	public void applyEffectsToFighterSpawnedByShip(ShipAPI fighter, ShipAPI ship, String id) {
-		super.applyEffectsToFighterSpawnedByShip(fighter, ship, id);
-	}
-
 	//Stats
     public static final float TIME_MULT = 1.2f;
     public static final Color AFTERIMAGE_COLOR_STANDARD = new Color(80, 255, 38, 100);
