+++ jars/src/data/scripts/weapons/VassRandomPrototypeScript.java
@@ -2,13 +2,10 @@ package data.scripts.weapons;
 
 import com.fs.starfarer.api.Global;
 import com.fs.starfarer.api.combat.*;
-import com.fs.starfarer.api.util.Misc;
 import data.scripts.campaign.barEvents.VassPerturbaWeaponTestingIntel;
-import data.scripts.utils.VassCyllelFarchogGuidanceScript;
-import org.jetbrains.annotations.Nullable;
+import data.scripts.utils.VassPerturbaRandomPrototypeManager.PrototypeWeaponData;
 import org.lazywizard.lazylib.MathUtils;
 import org.lazywizard.lazylib.combat.CombatUtils;
-import org.lwjgl.util.vector.Vector2f;
 
 import java.awt.*;
 import java.util.ArrayList;
@@ -20,12 +17,19 @@ import java.util.List;
  */
 public class VassRandomPrototypeScript implements EveryFrameWeaponEffectPlugin {
     private List<DamagingProjectileAPI> alreadyRegisteredProjectiles = new ArrayList<>();
+    private boolean hasReducedHealth = false;
 
     @Override
     public void advance(float amount, CombatEngineAPI engine, WeaponAPI weapon) {
         //Failsafe : refit screen and more!
         if (engine == null || !engine.isEntityInPlay(weapon.getShip())) {
             return;
+        }
+
+        //Prototype weapons always have half the normal weapon health
+        if (!hasReducedHealth) {
+            hasReducedHealth = true;
+            weapon.setCurrHealth(weapon.getCurrHealth()/2f);
         }
 
         //Instantly disable in simulation or missions
@@ -56,26 +60,32 @@ public class VassRandomPrototypeScript implements EveryFrameWeaponEffectPlugin {
         }
 
         for (DamagingProjectileAPI proj : CombatUtils.getProjectilesWithinRange(weapon.getLocation(), 200f)) {
-            if (proj.getProjectileSpecId().equals("vass_fake_prototype_shot") ||
+            //Projectile replacement! Only for our "initial" fake projectile, though
+            if (("vass_fake_prototype_shot").equals(proj.getProjectileSpecId()) &&
                     proj.getWeapon() == weapon && !alreadyRegisteredProjectiles.contains(proj)
                     && engine.isEntityInPlay(proj) && !proj.didDamage()) {
-                //Projectile replacement! Only for our "initial" fake projectile, though
-                DamagingProjectileAPI newProj = (DamagingProjectileAPI) engine.spawnProjectile(weapon.getShip(), weapon,
-                        currentData.projectileWeaponId, proj.getLocation(),
-                        proj.getFacing()+ MathUtils.getRandomNumberInRange(-currentData.inaccuracy, currentData.inaccuracy),
-                        null);
-                newProj.setDamageAmount(newProj.getDamageAmount()*currentData.damageMult);
-                float speedMultThisShot = 1f + MathUtils.getRandomNumberInRange(-1f, 1f) * currentData.speedVariation;
-                newProj.getVelocity().x = newProj.getVelocity().x * speedMultThisShot + weapon.getShip().getVelocity().x;
-                newProj.getVelocity().y = newProj.getVelocity().y * speedMultThisShot + weapon.getShip().getVelocity().y;
-                Global.getSoundPlayer().playSound(currentData.sound, 1f, 1f, proj.getLocation(), weapon.getShip().getVelocity());
-                engine.removeEntity(proj);
+                float shotgunningLeft = currentData.shotgunFactor;
+                while (Math.random() < shotgunningLeft) {
+                    shotgunningLeft--;
+                    DamagingProjectileAPI newProj = (DamagingProjectileAPI) engine.spawnProjectile(weapon.getShip(), weapon,
+                            currentData.projectileWeaponId, proj.getLocation(),
+                            proj.getFacing()+ MathUtils.getRandomNumberInRange(-currentData.inaccuracy, currentData.inaccuracy),
+                            weapon.getShip().getVelocity());
+                    newProj.setDamageAmount(newProj.getDamageAmount()*currentData.damageMult/currentData.shotgunFactor);
+                    float speedMultThisShot = 1f + MathUtils.getRandomNumberInRange(-1f, 1f) * currentData.speedVariation;
+                    newProj.getVelocity().x = newProj.getVelocity().x * speedMultThisShot + weapon.getShip().getVelocity().x;
+                    newProj.getVelocity().y = newProj.getVelocity().y * speedMultThisShot + weapon.getShip().getVelocity().y;
+                    Global.getSoundPlayer().playSound(currentData.sound, 1f, 1f, proj.getLocation(), weapon.getShip().getVelocity());
 
-                //We use guidance, if supplied
-                if (currentData.guidance != null) {
-                    currentData.guidance.applyToProjectile(newProj);
+                    //We use guidance, if supplied
+                    if (currentData.guidance != null) {
+                        currentData.guidance.applyToProjectile(newProj);
+                    }
+                    alreadyRegisteredProjectiles.add(newProj);
                 }
-                alreadyRegisteredProjectiles.add(newProj);
+
+                alreadyRegisteredProjectiles.add(proj);
+                engine.removeEntity(proj);
             }
         }
 
@@ -90,54 +100,5 @@ public class VassRandomPrototypeScript implements EveryFrameWeaponEffectPlugin {
         if (currentData.weaponEffectPlugin != null) {
             currentData.weaponEffectPlugin.advance(amount, engine, weapon);
         }
-    }
-
-    //Public class for storing prototype weapon data
-    public static class PrototypeWeaponData {
-        public final float damageMult;
-        public final float reloadMult;
-        public final float inaccuracy;
-        public final float speedVariation;
-        public final String projectileWeaponId;
-        public final boolean pd;
-        public final GuidanceApplier guidance;
-        public final EveryFrameWeaponEffectPlugin weaponEffectPlugin;
-        public final String sound;
-
-        /**
-         * Creates a new prototype weapon data with the given parameters
-         * @param damageMult damage multiplier of the shot
-         * @param reloadMult expected to be <= 1f: lower numbers mean faster firerate
-         * @param inaccuracy in degrees to each side: the projectile will ween off this much to either side when firing
-         * @param speedVariation how much an individual projectile may diverge from average projectile speed
-         * @param projectileWeaponId ID of the weapon to take projectiles from
-         * @param pd true if the weapon is to be treated as a PD weapon, false otherwise
-         * @param guidance guidance script to apply to each individual projectile: could of course be part of
-         *                 weaponEffectScript, but is here so it can be more easily randomized and parametrized
-         * @param weaponEffectPlugin arbitrary script to apply to the weapon as its everyFrameWeaponEffect.
-         *                           NOTE: expected to be "stateless" or track each weapon individually in a map
-         *                           as this is a shared script instance between all the prototypes
-         */
-        public PrototypeWeaponData(float damageMult, float reloadMult, float inaccuracy, float speedVariation,
-                                   String projectileWeaponId, boolean pd, @Nullable GuidanceApplier guidance,
-                                   @Nullable EveryFrameWeaponEffectPlugin weaponEffectPlugin, String sound) {
-            this.damageMult = damageMult;
-            this.reloadMult = reloadMult;
-            this.inaccuracy = inaccuracy;
-            this.speedVariation = speedVariation;
-            this.projectileWeaponId = projectileWeaponId;
-            this.pd = pd;
-            this.guidance = guidance;
-            this.weaponEffectPlugin = weaponEffectPlugin;
-            this.sound = sound;
-        }
-    }
-
-    /**
-     * Public class defining a simple guidance applying script for a randomized weapon: expected to
-     * apply a new MagicGuidedProjectileScript to the projectiles it recieves with somewhat consistant behaviour
-     */
-    public static abstract class GuidanceApplier {
-        public abstract void applyToProjectile(DamagingProjectileAPI proj);
     }
 }
